conda_requirements.txt
aom=3.7.1=h463b476_0
astropy=6.0.0=py310h50ce23c_0
astropy-iers-data=0.2023.12.18.0.30.18=pyhd8ed1ab_0
brotli=1.1.0=hb547adb_1
brotli-bin=1.1.0=hb547adb_1
brotli-python=1.1.0=py310h1253130_1
bzip2=1.0.8=h93a5062_5
c-ares=1.24.0=h93a5062_0
ca-certificates=2023.11.17=hf0a4a13_0
cairo=1.18.0=hd1e100b_0
cctools=973.0.1=h4faf515_15
cctools_osx-arm64=973.0.1=h62378fb_15
certifi=2023.11.17=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
clang=16.0.6=haab561b_3
clang-16=16.0.6=default_hd209bcb_3
clang_impl_osx-arm64=16.0.6=hc421ffc_7
clang_osx-arm64=16.0.6=h54d7cd3_7
clangxx=16.0.6=default_h5c94ee4_3
colorama=0.4.6=pyhd8ed1ab_0
compiler-rt=16.0.6=h3808999_2
compiler-rt_osx-arm64=16.0.6=h3808999_2
contextlib2=21.6.0=pyhd8ed1ab_0
contourpy=1.2.0=py310hd137fd4_0
cycler=0.12.1=pyhd8ed1ab_0
dav1d=1.2.1=hb547adb_0
exceptiongroup=1.2.0=pyhd8ed1ab_0
execnet=2.0.2=pyhd8ed1ab_0
expat=2.5.0=hb7217d7_1
ffmpeg=6.1.0=gpl_hfdc440f_105
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_1
fontconfig=2.14.2=h82840c6_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.47.0=py310hd125d64_0
freetype=2.12.1=hadb7bae_2
fribidi=1.0.10=h27ca646_0
fsspec=2023.12.2=pypi_0
gettext=0.21.1=h0186832_0
gfortran=13.2.0=h1ca8e4b_1
gfortran_impl_osx-arm64=13.2.0=h30f4408_1
gfortran_osx-arm64=13.2.0=h57527a5_1
gmp=6.3.0=h965bd2d_0
gnutls=3.7.9=hd26332c_0
graphite2=1.3.13=h9f76cd9_1001
harfbuzz=8.3.0=h8f0ba13_0
hdf5=1.14.3=nompi_h5bb55e9_100
icu=73.2=hc8870d7_0
idna=3.6=pyhd8ed1ab_0
importlib-metadata=7.0.0=pyha770c72_0
iniconfig=2.0.0=pyhd8ed1ab_0
isl=0.25=h9a09cb3_0
jasper=4.1.0=h7c0e182_0
jinja2=3.1.2=pypi_0
kiwisolver=1.4.5=py310h38f39d4_1
krb5=1.21.2=h92f50d5_0
lame=3.100=h1a8c8d9_1003
lcms2=2.16=ha0e7c42_0
ld64=609=h634c8be_15
ld64_osx-arm64=609=ha4bd21c_15
lerc=4.0.0=h9a09cb3_0
libabseil=20230802.1=cxx17_h13dd4ca_0
libaec=1.1.2=h13dd4ca_1
libass=0.17.1=hf7da4fe_1
libblas=3.9.0=20_osxarm64_openblas
libbrotlicommon=1.1.0=hb547adb_1
libbrotlidec=1.1.0=hb547adb_1
libbrotlienc=1.1.0=hb547adb_1
libcblas=3.9.0=20_osxarm64_openblas
libclang-cpp16=16.0.6=default_hd209bcb_3
libcurl=8.5.0=h2d989ff_0
libcxx=16.0.6=h4653b0c_0
libdeflate=1.19=hb547adb_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h93a5062_2
libexpat=2.5.0=hb7217d7_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_1
libgfortran-devel_osx-arm64=13.2.0=h5d7a38c_1
libgfortran5=13.2.0=hf226fd6_1
libglib=2.78.3=hb438215_0
libiconv=1.17=h0d3ecfb_2
libidn2=2.3.4=h1a8c8d9_0
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=20_osxarm64_openblas
liblapacke=3.9.0=20_osxarm64_openblas
libllvm16=16.0.6=haab561b_3
libnghttp2=1.58.0=ha4dd798_1
libopenblas=0.3.25=openmp_h6c19121_0
libopencv=4.8.1=py310he8a427e_5
libopenvino=2023.1.0=h965bd2d_2
libopenvino-arm-cpu-plugin=2023.1.0=h965bd2d_2
libopenvino-auto-batch-plugin=2023.1.0=h965bd2d_2
libopenvino-auto-plugin=2023.1.0=h965bd2d_2
libopenvino-hetero-plugin=2023.1.0=h965bd2d_2
libopenvino-ir-frontend=2023.1.0=h965bd2d_2
libopenvino-onnx-frontend=2023.1.0=h965bd2d_2
libopenvino-paddle-frontend=2023.1.0=h965bd2d_2
libopenvino-pytorch-frontend=2023.1.0=h965bd2d_2
libopenvino-tensorflow-frontend=2023.1.0=h965bd2d_2
libopenvino-tensorflow-lite-frontend=2023.1.0=h965bd2d_2
libopus=1.3.1=h27ca646_1
libpng=1.6.39=h76d750c_0
libprotobuf=4.24.4=hc9861d8_0
libsqlite=3.44.2=h091b4b1_0
libssh2=1.11.0=h7a5bd25_0
libtasn1=4.19.0=h1a8c8d9_0
libtiff=4.6.0=ha8a6c65_2
libunistring=0.9.10=h3422bc3_0
libvpx=1.13.1=hb765f3a_0
libwebp-base=1.3.2=hb547adb_0
libxcb=1.15=hf346824_0
libxml2=2.12.3=h0d0cfa8_0
libzlib=1.2.13=h53f4e23_5
llvm-openmp=17.0.6=hcd81f8e_0
llvm-tools=16.0.6=haab561b_3
markupsafe=2.1.3=pypi_0
matplotlib=3.8.2=py310hb6292c7_0
matplotlib-base=3.8.2=py310h9d2df84_0
mock=5.1.0=pyhd8ed1ab_0
mpc=1.3.1=h91ba8db_0
mpfr=4.2.1=h9546428_0
mpmath=1.3.0=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.4=h463b476_2
nettle=3.9.1=h40ed0f5_0
networkx=3.2.1=pypi_0
numpy=1.26.2=py310h30ee222_0
opencv=4.8.1=py310h66be37d_5
openh264=2.4.0=h965bd2d_0
openjpeg=2.5.0=h4c1507b_3
openssl=3.2.0=h0d3ecfb_1
p11-kit=0.24.1=h29577a5_0
packaging=23.2=pyhd8ed1ab_0
path=16.7.1=py310hbe9552e_1
path.py=12.5.0=0
pcre2=10.42=h26f9a81_0
pillow=10.1.0=py310hfae7ebd_0
pip=23.3.2=pyhd8ed1ab_0
pixman=0.42.2=h13dd4ca_0
pluggy=1.3.0=pyhd8ed1ab_0
pthread-stubs=0.4=h27ca646_1001
pugixml=1.14=h13dd4ca_0
py-opencv=4.8.1=py310h5c099be_5
pyerfa=2.0.1.1=py310h50ce23c_0
pyparsing=3.1.1=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
pytest=7.4.3=pyhd8ed1ab_0
pytest-shutil=1.7.0=pyhd8ed1ab_1
python=3.10.13=h2469fbe_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=1.0.1=pyhd8ed1ab_0
python_abi=3.10=4_cp310
pyyaml=6.0.1=py310h2aa6e3c_1
readline=8.2=h92ec313_1
requests=2.31.0=pyhd8ed1ab_0
scipy=1.11.4=py310h2b794db_0
setuptools=68.2.2=pyhd8ed1ab_0
sigtool=0.1.3=h44b9a77_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.10=h17c5cce_0
spacepy=0.4.1=pypi_0
svt-av1=1.8.0=h463b476_0
sympy=1.12=pypi_0
tapi=1100.0.11=he4954df_0
tbb=2021.10.0=h1995070_2
termcolor=2.3.0=pyhd8ed1ab_0
tk=8.6.13=h5083fa2_1
tomli=2.0.1=pyhd8ed1ab_0
torch=2.1.2=pypi_0
torchaudio=2.1.2=pypi_0
torchvision=0.16.2=pypi_0
tornado=6.3.3=py310h2aa6e3c_1
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023c=h71feb2d_0
unicodedata2=15.1.0=py310h2aa6e3c_0
urllib3=2.1.0=pyhd8ed1ab_0
wheel=0.42.0=pyhd8ed1ab_0
x264=1!164.3095=h57fd34a_2
x265=3.5=hbc6ce65_3
xorg-libxau=1.0.11=hb547adb_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h53f4e23_5
zstd=1.5.5=h4f39d0f_0